move configuration from ToolInstallation to GlobalConfiguration (#43)
* displayname of node in pipeline

properly get the node when BuildData is initilaized from a pipeline
pump mockito to latest version
add integration test, that use Jenkins test harness

* refactoring

the plugin used a ToolInstallation for configuration. But logstash is
not a tool that can be installed. Instead this should be a
GlobalConfiguration.
The indexers have different configuration requirements. Host and port
are common, but things like username are not required everywhere. Also
the previous implementation was totally unflexible. In order to add a
new indexer one has to modify the sources. By configuring the indexers
via an extension point we make it easily possible to have individual
configuration jelly for each type. Adding a new indexer is just adding
new classes. So it can be even moved to a separate plugin.

Many of the fields in the plugin were public or package private. They
have been made private or protected with corresponding getters and
setters and direct calls to the fields replaced.

Get the charset of a run and use it whenever we need to convert the
String to a byte array

Use DateFormatter from apache common lang to avoid potential threading
problems.

* jenkins and java

use Jenkins 2.7.4 instead of 2.60.3
fix problem with cast that is requried with java 7

* fix findbugs issues in buildData

Use FastDateFormat from apache which is thread safe

Avoid possible NPE in Node detection
Avoid possible NPE in result (previous logic already avoided it but findbugs
still claimed it to be a problem)

Node detection is same for pipeline and freestyle so move it to initData
method

* javadoc and readme

update javadoc to get more information for developers that want to
extend the plugin.
Update the readme
Use the not deprecated junit runner of mockito

* remove unnecessary dependency to workflow-step-api

* do the comparison if something has changed in the configuration class

* incorporate feedback

rename junit test methods
synchronize getting the dao instance to avoid problems.
deprecate enums

* adjust link to plugin page

* simplify equals

* host based indexers vs uri

introduce an abstract host based indexer that serves as base for most
indexers. Elastic search uses a uri approach where we need scheme, host,
port and url path, so we allow to configure a uri instead.

* add some todos

* use URL instead of URI

stapler is not able to autconvert String to URI
add more TODOs

* use URI as persistence for elastic search url

URI can be easier compared
some minor changes in tests

* fix test

* use ExpectedException rule