charset for rabbitMQ (#58)
We passed the charset from the build to the Dao and used it in
RabbitMQ. This is wrong.
The dao is used by multiple builds in parallel, if a new build uses a
different charset (pipeline uses UTF-8 while freestyle use the default
charset) we change it for running builds
RabbitMQ stores the data in binary form so we need a charset. But the
consumer needs to know which charset is used, so we should make this
configurable.