[JENKINS-52712] Make compatible with JEP-200 (#79)
* [JENKINS-52712] Make compatible with JEP-200

RabbitMQ stores Charset which is not in the whitelist for XStream and
remoting. So the configuration is not persisted to disk without getting
notified in the UI.
Changing Charset to String is a compatible change with respect to the
persistance.

* handle possible wrong charset settings

e.g. when it is manually changed in config xml to a non existing value this
cóuld have lead to an execption when creating the dao.
Now we fallback to the default (for backwards compatibility).